Partially revert fbc1b6f
Since that commit, setup has not worked, because setup loads client
which loads packets which prepares a statement for a table that does not
exist yet, as database migrations have not been run.

Alteratively, we could do migrations earlier in setup, but those would
send a lot of log lines to the terminal in the middle of interactive
prompting.